+++ CHANGELOG.md
@@ -8,7 +8,10 @@ Note to self: See the bottom of this file for the release template text.
  - Add `//dome+`, which allows you to change the direction the dome is pointing in, and also create multiple domes at once
  - Add `//metaball`, which renders 2 or more [metaballs](https://en.wikipedia.org/wiki/Metaballs) in Minetest
  - Migrate from `depends.txt` to `mod.conf`
- - `//sculpt`: Fix undefined `default` brush
+ - `//sculpt`:
+ 	- Fix undefined `default` brush
+	- Change defaults to `circle`, `height=1`, and `brushsize=8`.
+	- **TODO:** change ordering to put `height` after `brushsize` instead of the other way around
  - Commands that modify the terrain now ignore liquids
  - `//hollow`: Fix safe region bug
  - Significant backend refactoring to tidy things up